refactoring - controllers
